requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
httplib2
psycopg2-binary
bs4
lxml
PyTelegramBotAPI
